requirements.txt
torch
torchvision
tqdm
scikit-image
numpy
pillow
matplotlib